requirements/dev.txt
-r common.txt
pytest~=6.0.1
pytest-aiohttp~=0.3.0
pytest-cov~=2.10.0
